+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## UNRELEASED
 
 * [FEATURE] Implement shard retention based on Prometheus data retention (it requires the `PrometheusShardRetentionPolicy` feature gate). #8478
+* [ENHANCEMENT] Add `cipherSuites` support for Thanos Sidecars and Rulers. #8524
 * [BUGFIX] Ensure that inactive shards don't scrape any targets when the sharding retention policy is `Retain`. #8513
 
 ## 0.90.1 / 2026-03-25